+++ src/DEM/DEMBase.cc
@@ -18,6 +18,7 @@
 #include "DataBase/ReplaceState.hh"
 #include "DataBase/DataBase.hh"
 
+
 #include "Field/FieldList.hh"
 #include "Field/NodeIterators.hh"