+++ src/Hydro/PressurePolicy.cc
@@ -56,7 +56,7 @@ update(const KeyType& key,
   KeyType fieldKey, nodeListKey;
   StateBase<Dimension>::splitFieldKey(key, fieldKey, nodeListKey);
   REQUIRE((fieldKey == HydroFieldNames::pressure or 
-           fieldKey == FSIFieldNames::rawPressure));
+           fieldKey == FSIFieldNames::damagedPressure));
   auto& P = state.field(key, Scalar());
 
   // Get the eos.  This cast is ugly, but is a work-around for now.
@@ -97,7 +97,16 @@ update(const KeyType& key,
 
   }
 
+  // Is someone trying to keep the damaged pressure in an independent Field?
+  // (I'm looking at you FSISPH)
+  const auto separateDamage = state.registered(buildKey(FSIFieldNames::damagedPressure));
+  Field<Dimension, Scalar>* PdPtr = nullptr;
+  if (separateDamage) PdPtr = &state.field(buildKey(FSIFieldNames::damagedPressure), 0.0);
+
   // If there's damage for this material, apply it to the pressure
+  // This is complicated by FSISPH, which wants to keep track of the damaged pressure separately,
+  // so we check if someone has registered damaged pressure as a field for this NodeList, in
+  // which case we apply the damage to a new copy of the pressure.
   if (state.registered(buildKey(SolidFieldNames::tensorDamage))) {
     const auto& D = state.field(buildKey(SolidFieldNames::tensorDamage), SymTensor::zero);
 
@@ -114,7 +123,12 @@ update(const KeyType& key,
     for (auto i = 0u; i < ni; ++i) {
       const auto Di = std::max(0.0, std::min(1.0, D(i).eigenValues().maxElement()));
       CHECK(Di >= 0.0 and Di <= 1.0);
-      P(i) = std::max(Pmin, (1.0 - Di)*P(i)) + std::max(PminDamage, Di*P(i));
+      const auto Pdi = std::max(Pmin, (1.0 - Di)*P(i)) + std::max(PminDamage, Di*P(i));
+      if (separateDamage) {
+        (*PdPtr)(i) = Pdi;
+      } else {
+        P(i) = Pdi;
+      }
     }
   }
 
@@ -122,6 +136,7 @@ update(const KeyType& key,
   if (usePorosity) {
     const auto& alpha = state.field(buildKey(SolidFieldNames::porosityAlpha), 0.0);
     P /= alpha;
+    if (separateDamage) (*PdPtr) /= alpha;
   }
 }